pyproject.toml
[project]
name = "bass-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "accelerate>=1.3.0",
    "audioldm>=0.0.10",
    "datasets>=3.3.2",
    "diffusers>=0.32.2",
    "discord-py>=2.4.0",
    "einops>=0.8.1",
    "huggingface-hub>=0.28.1",
    "jupyter>=1.1.1",
    "librosa>=0.10.2.post1",
    "notebook>=7.3.2",
    "pandas>=2.1.0",
    "python-dotenv>=1.0.1",
    "sentence-transformers>=3.4.1",
    "torchaudio>=2.6.0",
    "transformers>=4.48.2",
    "wandb>=0.19.7",
]
